requirements.txt
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.47
langchain_experimental==0.3.4
langchain_groq==0.3.1
langgraph==0.3.18
matplotlib==3.7.5
python-dotenv==1.0.1
streamlit==1.41.1
